% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workshop-functions.R
\name{conduct.survey}
\alias{conduct.survey}
\title{Conduct a survey to collect acoustic detection data}
\usage{
conduct.survey(skip.wait = FALSE)
}
\arguments{
\item{skip.wait}{For development purposes only. Don't set to
\code{TRUE}!}
}
\value{
A list with two components: \code{captures} includes the
detection data, and \code{traps} includes the listening post
locations.
}
\description{
Creates a plot of Phnom Prich Wildlife Sanctuary and invites the
user to select locations at which to deploy listening posts.
}
