% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workshop-functions.R
\name{interpolate.covs}
\alias{interpolate.covs}
\title{Interpolate a spatial covariate}
\usage{
interpolate.covs(df)
}
\arguments{
\item{df}{A data frame with measured covariate values.}
}
\value{
A data frame. The first two columns are x- and
y-coordinates covering Phnom Prich Wildlife
Sanctuary. Remaining columns provide interpolated covariate
values.
}
\description{
Conducts spatial interpolation to impute values of covariates
throughout Phnom Prich Wildlife Sanctuary.
}
