% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workshop-functions.R
\name{measure.covariates}
\alias{measure.covariates}
\title{Conduct a survey to measure spatial covariates}
\usage{
measure.covariates(skip.wait = FALSE)
}
\arguments{
\item{skip.wait}{For development purposes only. Don't set to
\code{TRUE}!}
}
\value{
A data frame with columns for the coordinates of the
selected locations, and a column for each observed spatial
variable.
}
\description{
Creates a plot of Phnom Prich Wildlife Sanctuary and invites the
user to select sampling locations to measure covariates.
}
