% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workshop-functions.R
\name{plotcov}
\alias{plotcov}
\title{Plot a spatial covariate}
\usage{
plotcov(df, cov.name)
}
\arguments{
\item{df}{A data frame with measured or interpolated covariate
values.}

\item{cov.name}{A character string, indicating which column in
\code{df} to plot.}
}
\description{
Plots a spatial covariate at the points provided in the first
argument.
}
