% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workshop-functions.R
\docType{data}
\name{ppws}
\alias{ppws}
\title{A grid of cells covering Phnom Prich Wildlife Sanctuary}
\format{
\subsection{\code{ppws}}{

A data frame with two columns: x and y.
}
}
\usage{
ppws
}
\description{
A data frame specifying x- and y-coordinates for a grid of cells
that covers Phnom Prich Wildlife Sanctuary.
}
\keyword{datasets}
