% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workshop-functions.R
\docType{data}
\name{villages.df}
\alias{villages.df}
\title{Locations of villages in Phnom Prich Wildlife Sanctuary}
\format{
\subsection{\code{ppws}}{

A data frame with two columns: x and y.
}
}
\usage{
villages.df
}
\description{
A data frame specifying x- and y-coordinates for villages.
}
\keyword{datasets}
